webrtc: avoid blocking the PeerConnection thread
webrtcbin has a thread in which all signals are dispatched from. To
avoid blocking that thread, first yield to allow the blocking queued
call to return.

Care has to be taken if a signal results in a new JS wrapper, as it has
to be created before we yield or we risk loosing events.